pyproject.toml
[tool.poetry]
name = "drf-online-education"
version = "0.1.0"
description = ""
authors = ["egor <ega.nik5674@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
redis = "^5.0.4"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2024.2"
ipython = "^8.24.0"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
